wip: root sidebar
